battle-giveaway/requirements.txt
instaloader>=4.14,<5
Pillow>=11.1.0,<12
requests>=2.32.0,<3
